requirements.txt
Flask==3.0.0
requests==2.31.0
youtube-transcript-api==0.6.1
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Flask-Limiter==3.5.0
PyJWT==2.8.0

